Add files via upload
This is the bot python file, it looks at each professors which are teaching that course for the semester, selects the best according to the USF Infocenter grades data (highest class avg gpa). It also handles time conflicts by prioritizing the courses entered first, meaning if there is a time conflict between the course entered first and the course entered second, it will select the best prof (by avg gpa from infocenter) and then it will select the best non-conflicting prof from course 2 and this continues until the last course. It finally displays a proposed plan.